move menu_draw into render_draw
