+++ src/render_draw.c
@@ -6,11 +6,48 @@
 /*   By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/20 15:52:55 by gwolf             #+#    #+#             */
-/*   Updated: 2024/01/03 12:04:11 by gwolf            ###   ########.fr       */
+/*   Updated: 2024/01/06 12:49:14 by gwolf            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "render.h"
+
+static uint32_t	fast_alpha_blend(uint32_t bg_color, t_menu menu)
+{
+	uint32_t	bg_rb;
+	uint32_t	bg_g;
+	uint32_t	blend_rb;
+	uint32_t	blend_g;
+
+	bg_rb = (bg_color & 0xFF00FF) * menu.inv_alpha;
+	bg_g = (bg_color & 0x00FF00) * menu.inv_alpha;
+	blend_rb = (((menu.rb + bg_rb) >> 8) & 0xFF00FF);
+	blend_g = (((menu.g + bg_g) >> 8) & 0x00FF00);
+	return (blend_rb | blend_g);
+}
+
+void	ft_blend_background(t_img *img, t_img *veil, t_menu menu)
+{
+	char		*img_pixel;
+	char		*veil_pixel;
+	uint32_t	blend_colour;
+	t_vec2i		pos;
+
+	pos.y = -1;
+	while (++pos.y < veil->size.y)
+	{
+		pos.x = -1;
+		while (++pos.x < veil->size.x)
+		{
+			img_pixel = img->addr
+				+ (pos.y * img->line_len + pos.x * img->bytes);
+			blend_colour = fast_alpha_blend(*(uint32_t *)img_pixel, menu);
+			veil_pixel = veil->addr
+				+ (pos.y * veil->line_len + pos.x * veil->bytes);
+			*(uint32_t *)veil_pixel = blend_colour;
+		}
+	}
+}
 
 void	ft_put_pix_to_image(t_img *img, int x, int y, int color)
 {
